requirements.txt
yfinance
pandas
numpy
newsapi-python
python-dotenv